Use size_t for size containing variables
Sizes should not be represented as uint32_t as
this is non portable on 64 bit architectures.
It also raises warnings when narrowing size_t to uint32_t
on 64 bit architectures.

Change all size variables to be size_t.
The only exception is nework code as the protocol
should not be architecture dependent.

+++ parameter/BitParameterBlock.h
@@ -37,13 +37,13 @@ public:
     CBitParameterBlock(const std::string& strName, const CTypeElement* pTypeElement);
 
     // Instantiation, allocation
-    virtual uint32_t getFootPrint() const;
+    virtual size_t getFootPrint() const;
 
     // Type
     virtual Type getType() const;
 
     // Size
-    uint32_t getSize() const;
+    size_t getSize() const;
 
     // Used for simulation and virtual subsystems
     virtual void setDefaultValues(CParameterAccessContext& parameterAccessContext) const;
